+++ src/types.js
@@ -1300,8 +1300,14 @@ class ResourceNode {
    * @param {string} fhirNodeDataType - FHIR node data type, if the resource node
    *  is described in the FHIR model.
    *  @param {Object} model - the model object specific to a domain, e.g. R4.
+   *  @param {string} propName – the property name of the node within its parent
+   *   resource, if applicable. Used to identify the specific property this node
+   *   represents.
+   *  @param {number} index - the index of the node within an array property,
+   *   if the node is part of an array. Used to track the position of the node
+   *   among siblings.
    */
-  constructor(data, parentResNode, path, _data, fhirNodeDataType, model) {
+  constructor(data, parentResNode, path, _data, fhirNodeDataType, model, propName = null, index = null) {
     // If data is a resource (maybe a contained resource) reset the path
     // information to the resource type.
     if (data?.resourceType) {
@@ -1314,6 +1320,8 @@ class ResourceNode {
     this._data = _data || {};
     this.fhirNodeDataType = fhirNodeDataType || null;
     this.model = model || null;
+    this.propName = propName;
+    this.index = index;
   }
 
   /**
@@ -1384,6 +1392,22 @@ class ResourceNode {
     return this.convertedData;
   }
 
+  /**
+   * The full property name of the node in the resource
+   * (e.g. Patient.name[0].given[0])
+   *
+   * @return {string} - returns the full property name of the node in
+   *  the resource.
+   */
+  fullPropertyName() {
+    let result = this.parentResNode ?
+      this.parentResNode.fullPropertyName() + '.' + this.propName
+      : this.propName || this.path;
+    if (this.index != null) {
+      result += '[' + this.index + ']';
+    }
+    return result;
+  }
 }
 
 
@@ -1392,9 +1416,12 @@ class ResourceNode {
  *  given node is already a ResourceNode.  Takes the same arguments as the
  *  constructor for ResourceNode.
  */
-ResourceNode.makeResNode = function(data, parentResNode, path, _data, fhirNodeDataType, model) {
-  return (data instanceof ResourceNode) ? data : new ResourceNode(data, parentResNode, path, _data, fhirNodeDataType, model);
+ResourceNode.makeResNode = function(data, parentResNode, path, _data, fhirNodeDataType, model, propName, index) {
+  return (data instanceof ResourceNode) ? data
+    : new ResourceNode(data, parentResNode, path, _data, fhirNodeDataType,
+      model, propName, index);
 };
+
 
 // The set of available data types in the System namespace
 const availableSystemTypes = new Set();